Create the :RenameDirs command
This is the first commit of my replacer.vim plugin. It introduces the
RenameDirs function and command which will be used to find and replace
parts of a subdirectory name in vim.

This uses the `find` command in shell script because I have run into
performance issues when using other find and replace plugins that depend
on vim's grep search. As such, this plugin will only work on UNIX
machines.