Add vimdocs
To avoid duplication, I think it makes sense to convert the README file
into vimdocs for the plugin. The md2vim go package [1] does this quite
well but requires the removal of the Table of Contents already generated
by the markdown-toc plugin.[2] This commit, introduces a script that
will remove the README TOC before generating the formatted vimdoc.

[1]: https://github.com/FooSoft/md2vim
[2]: https://github.com/mzlogin/vim-markdown-toc